Fun: Finally fix animated gifs
This issue has existed since the very beginning of this module in
Rubbergoddess (and Rubbergod) codebase. Fairly recently (April 2022)
the issue of flickering GIFs has been fixed in Pillow 9.1.0 via PR 6128.

As the issue depends on the avatar of the user, it is not possible to
claim this fixes it 100 %. I do believe/hope it does.